Factor MarchingCubesSingleDomain into its own file.
