+++ src/axom/quest/MarchingCubes.hpp
@@ -72,8 +72,7 @@ class MarchingCubesSingleDomain;
  * cubes).
  *
  * The input mesh is a Conduit::Node following the Mesh Blueprint
- * convention.  The mesh must be in multi-domain format.  For
- * single-domain, see MarchingCubesSingleDomain.
+ * convention.  The mesh must be in multi-domain format.
  *
  * Usage example:
  * @beginverbatim
@@ -238,10 +237,9 @@ public:
            use the id set in the constructor.
 
     The buffer size is getContourCellCount().
-
-    Memory space of data corresponds to allocator set in the constructor.
   */
-  axom::Array<MarchingCubes::DomainIdType> getContourFacetDomainIds(
+  template<typename DomainIdType>
+  axom::Array<DomainIdType> getContourFacetDomainIds(
     int allocatorID = axom::INVALID_ALLOCATOR_ID) const;
 
   #if 1
@@ -296,7 +294,7 @@ private:
 
     May be longer than m_domainCount (the real count).
   */
-  axom::Array<std::unique_ptr<MarchingCubesSingleDomain>> m_singles;
+  axom::Array<std::shared_ptr<MarchingCubesSingleDomain>> m_singles;
   std::string m_topologyName;
   std::string m_fcnFieldName;
   std::string m_fcnPath;
@@ -333,225 +331,6 @@ private:
   //!@brief Allocate output buffers corresponding to runtime policy.
   void allocateOutputBuffers();
 };
-
-/*!
- * \@brief Class implementing marching cubes algorithm for a single
- *  domain.
- *
- * \sa MarchingCubes
- */
-class MarchingCubesSingleDomain
-{
-public:
-  using RuntimePolicy = axom::runtime_policy::Policy;
-  /*!
-   * \brief Constructor for applying algorithm in a single domain.
-   * See MarchingCubes for the multi-domain implementation.
-   *
-   * \param [in] runtimePolicy A value from RuntimePolicy.
-   *             The simplest policy is RuntimePolicy::seq, which specifies
-   *             running sequentially on the CPU.
-   * \param [in] allocatorID Data allocator ID.  Choose something compatible
-   *             with \c runtimePolicy.  See \c esecution_space.
-   * \param [in] dataPar Choice of data-parallel implementation.
-   */
-  MarchingCubesSingleDomain(RuntimePolicy runtimePolicy,
-                            int allocatorID,
-                            MarchingCubesDataParallelism dataPar);
-
-  /*!
-    @brief Intitialize object to a domain.
-    \param [in] dom Blueprint single-domain mesh containing scalar field.
-    \param [in] topologyName Name of Blueprint topology to use in \a dom
-    \param [in] maskField Cell-based std::int32_t mask field.  If provided,
-                cells where this field evaluates to false are skipped.
-
-    Array data in \a dom must be accessible in the the \a
-    runtimePolicy environment in the constructor.  It's an error if
-    not, e.g., using CPU memory with a GPU policy.
-
-    Some data from \a dom may be cached by the constructor.  Any
-    change to it without re-initialization leads to undefined
-    behavior.
-
-    The mesh coordinates should be stored contiguously.  See
-    conduit::blueprint::is_contiguous().  In the future, this
-    requirement may be relaxed, possibly at the cost of a
-    transformation and storage of the temporary contiguous layout.
-  */
-  void initialize( const conduit::Node &dom,
-                   const std::string &topologyName,
-                   const std::string &maskfield);
-
-  int spatialDimension() const { return m_ndim; }
-
-  /*!
-    @brief Specify the field containing the nodal scalar function
-    in the input mesh.
-    \param [in] fcnField Name of node-based scalar function values.
-  */
-  void setFunctionField(const std::string &fcnField)
-  {
-    m_fcnFieldName = fcnField;
-    m_fcnPath = "fields/" + fcnField;
-    SLIC_ASSERT(m_dom->has_path(m_fcnPath));
-    SLIC_ASSERT(m_dom->fetch_existing(m_fcnPath + "/association").as_string() ==
-                "vertex");
-    SLIC_ASSERT(m_dom->has_path(m_fcnPath + "/values"));
-    m_impl->setFunctionField(fcnField);
-  }
-
-  void setContourValue(double contourVal)
-  {
-    m_contourVal = contourVal;
-    if(m_impl) m_impl->setContourValue(m_contourVal);
-  }
-
-  // Methods trivially delegated to implementation.
-  void markCrossings() { m_impl->markCrossings(); }
-  void scanCrossings() { m_impl->scanCrossings(); }
-  axom::IndexType getContourCellCount()
-  {
-    return m_impl->getContourCellCount();
-  }
-  void computeContour() { m_impl->computeContour(); }
-
-  /*!
-    @brief Get the Blueprint domain id specified in \a state/domain_id
-    if it is provided, or use the given default if not provided.
-  */
-  MarchingCubes::DomainIdType getDomainId(MarchingCubes::DomainIdType defaultId) const;
-
-  //!@brief Get number of cells in the generated contour mesh.
-  axom::IndexType getContourCellCount() const
-  {
-    SLIC_ASSERT_MSG(
-      m_impl,
-      "There is no contour mesh until you call computeIsocontour()");
-    axom::IndexType cellCount = m_impl->getContourCellCount();
-    return cellCount;
-  }
-
-  //!@brief Get number of nodes in the generated contour mesh.
-  axom::IndexType getContourNodeCount() const
-  {
-    return m_ndim * getContourCellCount();
-  }
-
-  /*!
-    @brief Base class for implementations templated on dimension DIM
-    and execution space ExecSpace.
-
-    Implementation details templated on DIM and ExecSpace cannot
-    be in MarchingCubesSingleDomain so should live in this class.
-
-    This class allows m_impl to refer to any implementation used
-    at runtime.
-  */
-  struct ImplBase
-  {
-    /*!
-      @brief Prepare internal data for operating on the given domain.
-
-      Put in here codes that can't be in MarchingCubesSingleDomain
-      due to template use (DIM and ExecSpace).
-    */
-    virtual void initialize(const conduit::Node &dom,
-                            const std::string &topologyName,
-                            const std::string &maskPath = {}) = 0;
-
-    virtual void setFunctionField(const std::string &fcnFieldName) = 0;
-    virtual void setContourValue(double contourVal) = 0;
-
-    void setDataParallelism(MarchingCubesDataParallelism dataPar)
-    {
-      m_dataParallelism = dataPar;
-    }
-
-    //@{
-    //!@name Distinct phases in contour generation.
-    //!@brief Compute the contour mesh.
-    //!@brief Mark parent cells that cross the contour value.
-    virtual void markCrossings() = 0;
-    //!@brief Scan operations to determine counts and offsets.
-    virtual void scanCrossings() = 0;
-    //!@brief Compute contour data.
-    virtual void computeContour() = 0;
-    //@}
-
-    //@{
-    //!@name Output methods
-    //!@brief Return number of contour mesh facets generated.
-    virtual axom::IndexType getContourCellCount() const = 0;
-    //@}
-
-    void setOutputBuffers(axom::ArrayView<axom::IndexType, 2> &facetNodeIds,
-                          axom::ArrayView<double, 2> &facetNodeCoords,
-                          axom::ArrayView<axom::IndexType, 1> &facetParentIds,
-                          axom::IndexType facetIndexOffset)
-    {
-      m_facetNodeIds = facetNodeIds;
-      m_facetNodeCoords = facetNodeCoords;
-      m_facetParentIds = facetParentIds;
-      m_facetIndexOffset = facetIndexOffset;
-    }
-
-    virtual ~ImplBase() { }
-
-    virtual void clear() = 0;
-
-    MarchingCubesDataParallelism m_dataParallelism =
-      MarchingCubesDataParallelism::byPolicy;
-
-    double m_contourVal = 0.0;
-    axom::ArrayView<axom::IndexType, 2> m_facetNodeIds;
-    axom::ArrayView<double, 2> m_facetNodeCoords;
-    axom::ArrayView<IndexType> m_facetParentIds;
-    axom::IndexType m_facetIndexOffset = -1;
-  };
-
-  ImplBase& getImpl()
-  {
-    return *m_impl;
-  }
-
-private:
-  RuntimePolicy m_runtimePolicy;
-  int m_allocatorID = axom::INVALID_ALLOCATOR_ID;
-
-  //@brief Choice of full or partial data-parallelism, or byPolicy.
-  MarchingCubesDataParallelism m_dataParallelism =
-    MarchingCubesDataParallelism::byPolicy;
-
-  /*!
-    \brief Computational mesh as a conduit::Node.
-  */
-  const conduit::Node *m_dom;
-  int m_ndim;
-
-  //!@brief Name of Blueprint topology in m_dom.
-  std::string m_topologyName;
-
-  std::string m_fcnFieldName;
-  //!@brief Path to nodal scalar function in m_dom.
-  std::string m_fcnPath;
-
-  std::string m_maskFieldName;
-  //!@brief Path to mask in m_dom.
-  std::string m_maskPath;
-
-  double m_contourVal = 0.0;
-
-  std::unique_ptr<ImplBase> m_impl;
-
-  /*!
-   * \brief Set the blueprint single-domain mesh.
-   *
-   * Some data from \a dom may be cached.
-   */
-  void setDomain(const conduit::Node &dom);
-
-};  // class MarchingCubesSingleDomain
 
 }  // namespace quest
 }  // namespace axom
